Moved components and pages to app folder
