+++ nefac-website/src/app/pages/nefac-news/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Bubble from './../../components/Bubble';
+import Bubble from '../../components/Bubble';
 
 export const NewsPage: React.FC = () => {
     return (
